change int by size_t in 0

+++ 0x13-more_singly_linked_lists/0-print_listint.c
@@ -8,7 +8,7 @@
 
 size_t print_listint(const listint_t *h)
 {
-	int iterator = 0;
+	size_t iterator = 0;
 
 	if(!h)
 		return (0);
